refinement

+++ 2018/day_08.cpp
@@ -3,13 +3,14 @@
 #include "../utils/input.hpp"
 
 
-std::tuple<int, int, std::vector<int>>
-parse_data(const std::vector<int>& data) {
-    int num_children = data[0];
-    int num_meta = data[1];
-    std::vector<int> tail = std::vector<int>(data.begin() + 2, data.end());
-    std::vector<int> scores;
-    int meta_sum{0};
+std::tuple<size_t, size_t, std::vector<size_t>>
+parse_data(const std::vector<size_t>& data) {
+    size_t num_children = data[0];
+    size_t num_meta = data[1];
+    std::vector<size_t> tail
+      = std::vector<size_t>(data.begin() + 2, data.end());
+    std::vector<size_t> scores;
+    size_t meta_sum{0};
 
     for (size_t child_index{0}; child_index < num_children; child_index++) {
         auto update = parse_data(tail);
@@ -32,18 +33,18 @@ parse_data(const std::vector<int>& data) {
         }
     }
 
-    tail = std::vector<int>(tail.begin() + num_meta, tail.end());
+    tail = std::vector<size_t>(tail.begin() + num_meta, tail.end());
     return {meta_sum, score, tail};
 }
 
 
-int part_one(const std::vector<int>& input) {
+int part_one(const std::vector<size_t>& input) {
     auto [meta_sum, score, _] = parse_data(input);
     return meta_sum;
 }
 
 
-int part_two(const std::vector<int>& input) {
+int part_two(const std::vector<size_t>& input) {
     auto [meta_sum, score, _] = parse_data(input);
     return score;
 }
@@ -52,11 +53,11 @@ int part_two(const std::vector<int>& input) {
 int main() {
   utils::Reader reader(std::filesystem::path("../2018/data/input_08.txt"));
   auto input = utils::split_string(reader.get_lines()[0], ' ');
-  std::vector<int> transformed_input;
+  std::vector<size_t> transformed_input;
   std::transform(input.begin(),
                  input.end(),
                  std::back_insert_iterator(transformed_input),
-                 [] (const std::string& entry) { return std::stoi(entry); });
+                 [] (const std::string& entry) { return std::stoul(entry); });
 
   auto answer_one =  part_one(transformed_input);
   std::cout << "The answer to part one is: " << answer_one << std::endl;
